triggered clippy ICE \o/
